Work on Styling
Taking a different approach to this project. I really want it to be a gorgeous "Review Card" You could slot into a website somewhere. While the HTML is going to be eventually created with JS instead I got a template for what I want the final product to resemble, but the style isnt done yet.